can we go down to uint16_t?

+++ firmware/controllers/algo/obd_error_codes.h
@@ -12,7 +12,7 @@
 
 #include <cstdint>
 
-enum class ObdCode : uint32_t {
+enum class ObdCode : uint16_t {
 	None = 0,
 //P0001 Fuel Volume Regulator Control Circuit/Open
 //P0002 Fuel Volume Regulator Control Circuit Range/Performance
